requirements.txt
discord.py==2.1.0
pandas==1.5.2
pandas_market_calendars==4.1.3
python-dotenv==0.21.0
yfinance==0.2.3
